pyproject.toml
[project]
name = "attack-kg"
version = "2.0.0"
description = "Security remediation engine powered by MITRE ATT&CK"
requires-python = ">=3.11,<3.14"
dependencies = [
    "pyoxigraph>=0.3.22",
    "chromadb>=0.4.22",
    "sentence-transformers>=2.2.2",
    "einops>=0.7.0",
    "torch>=2.0.0",
    "ollama>=0.1.6",
    "openai>=1.12.0",
    "httpx>=0.26.0",
    "pydantic>=2.5.0",
    "typer>=0.9.0",
    "rich>=13.7.0",
    "rank-bm25>=0.2.2",
    "pyyaml>=6.0",
    "toon-format>=0.9.0b1",
]

[project.scripts]
attack-kg = "src.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[dependency-groups]
dev = ["pytest>=9.0.0"]
